+++ src/app/item-popup/SocketDetailsSelectedPlug.tsx
@@ -1,3 +1,4 @@
+import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import BungieImage from 'app/dim-ui/BungieImage';
 import { t } from 'app/i18next-t';
 import { canInsertPlug, insertPlug } from 'app/inventory/advanced-write-actions';
@@ -11,7 +12,14 @@ import {
 import { interpolateStatValue } from 'app/inventory/store/stats';
 import { destiny2CoreSettingsSelector, useD2Definitions } from 'app/manifest/selectors';
 import { showNotification } from 'app/notifications/notifications';
-import { DEFAULT_ORNAMENTS, DEFAULT_SHADER } from 'app/search/d2-known-values';
+import {
+  DEFAULT_ASPECTS,
+  DEFAULT_FRAGMENT,
+  DEFAULT_ORNAMENTS,
+  DEFAULT_PROJECTION,
+  DEFAULT_SHADER,
+  DEFAULT_TRANSMAT,
+} from 'app/search/d2-known-values';
 import { refreshIcon } from 'app/shell/icons';
 import AppIcon from 'app/shell/icons/AppIcon';
 import { useThunkDispatch } from 'app/store/thunk-dispatch';
@@ -20,7 +28,8 @@ import {
   isPlugStatActive,
   itemIsInstanced,
 } from 'app/utils/item-utils';
-import { StatHashes } from 'data/d2/generated-enums';
+import { DestinyItemSocketEntryDefinition } from 'bungie-api-ts/destiny2';
+import { SocketCategoryHashes, StatHashes } from 'data/d2/generated-enums';
 import { motion } from 'framer-motion';
 import _ from 'lodash';
 import React, { useState } from 'react';
@@ -36,6 +45,36 @@ const costStatHashes = [
   StatHashes.SolarCost,
   StatHashes.ArcCost,
 ];
+
+/** Figures out what kind of socket this is so that the "Apply" button can name the correct thing
+ * instead of generically saying "Select/Insert Mod".
+ * Note that these are used as part of the localization key.
+ */
+function uiCategorizeSocket(defs: D2ManifestDefinitions, socket: DestinyItemSocketEntryDefinition) {
+  const socketTypeDef = socket.socketTypeHash && defs.SocketType.get(socket.socketTypeHash);
+  if (socketTypeDef && socketTypeDef.socketCategoryHash === SocketCategoryHashes.Abilities) {
+    return 'Ability';
+  }
+
+  const initialPlugHash = socket.singleInitialItemHash;
+  if (initialPlugHash) {
+    if (initialPlugHash === DEFAULT_SHADER) {
+      return 'Shader';
+    } else if (DEFAULT_ORNAMENTS.includes(initialPlugHash)) {
+      return 'Ornament';
+    } else if (initialPlugHash === DEFAULT_FRAGMENT) {
+      return 'Fragment';
+    } else if (DEFAULT_ASPECTS.includes(initialPlugHash)) {
+      return 'Aspect';
+    } else if (initialPlugHash === DEFAULT_PROJECTION) {
+      return 'Projection';
+    } else if (initialPlugHash === DEFAULT_TRANSMAT) {
+      return 'Transmat';
+    }
+  }
+
+  return 'Mod';
+}
 
 export default function SocketDetailsSelectedPlug({
   plug,
@@ -113,19 +152,12 @@ export default function SocketDetailsSelectedPlug({
   const canDoAWA =
     itemIsInstanced(item) && canInsertPlug(socket, plug.hash, destiny2CoreSettings, defs);
 
-  const initialPlugHash = socket.socketDefinition.singleInitialItemHash;
-  let insertName;
-  if (initialPlugHash) {
-    if (initialPlugHash === DEFAULT_SHADER) {
-      insertName = canDoAWA ? t('Sockets.InsertShaderButton') : t('Sockets.SelectShaderButton');
-    } else if (DEFAULT_ORNAMENTS.includes(initialPlugHash)) {
-      insertName = canDoAWA ? t('Sockets.InsertOrnamentButton') : t('Sockets.SelectOrnamentButton');
-    } else {
-      insertName = canDoAWA ? t('Sockets.InsertModButton') : t('Sockets.SelectModButton');
-    }
-  } else {
-    insertName = canDoAWA ? t('Sockets.InsertModButton') : t('Sockets.SelectModButton');
-  }
+  // t(`Sockets.Insert`, { contextList: 'sockets' })
+  // t(`Sockets.Select`, { contextList: 'sockets' })
+  const kind = uiCategorizeSocket(defs, socket.socketDefinition);
+  const insertName = canDoAWA
+    ? t(`Sockets.Insert${kind}`, { contextList: 'sockets' })
+    : t(`Sockets.Select${kind}`, { contextList: 'sockets' });
 
   const [insertInProgress, setInsertInProgress] = useState(false);
   const onInsertPlug = async () => {