+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Removed notification when loading/updating a wish list. Go to the wish list section of the settings menu if you want to see details.
 * The progress notification for applying a loadout now shows each item and mod as it's being applied.
+* Mod picker now correctly names inserted thing (e.g Fragment, Shader).
 
 ## 6.97.1 <span class="changelog-date">(2021-12-26)</span>